InstrumentEditor: split into separate classes
Major refactoring of the instrument editor.

- `InstrumentEditor` does now only hold things concerned with the general instrument properties affecting all components (the general tab)
- The content of the `Layers` tab was moved into `ComponentView`
- The "Layers" tab was renamed into "Components"
- `ComponentsEditor` was introduced which is intended to hold an instance of `ComponentView` for each `InstrumentComponent` within an `Instrument`
- `InstrumentEditorPanel::m_nSelectedComponent` was moved into `ComponentsEditor`
- `InstrumentEditorPanel::m_nSelectedLayer` was moved into `ComponentView`

The big plan is to have something like a foldable tree in `ComponentsEditor`. When folded, the user can still see the name of the component, its solo and mute button as well as its gain. When unfolded, the previous content of the "Layers" tab will be shown for that particular `InstrumentComponent`
