+++ data/i18n/hydrogen_ca.ts
@@ -1389,6 +1389,55 @@ Please set your system&apos;s locale to UTF-8!</source>
     </message>
 </context>
 <context>
+    <name>ComponentView</name>
+    <message>
+        <source>Layer gain</source>
+        <translation>Guany de capa</translation>
+    </message>
+    <message>
+        <source>Component volume</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Layer pitch (Coarse)</source>
+        <translation>Ajust tò de capa (Fort)</translation>
+    </message>
+    <message>
+        <source>Layer pitch (Fine)</source>
+        <translation>Ajust tò de capa (Suau)</translation>
+    </message>
+    <message>
+        <source>Select selection algorithm</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>First in Velocity</source>
+        <extracomment>Sample selection algorithm available in the instrument editor</extracomment>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Round Robin</source>
+        <extracomment>Sample selection algorithm available in the instrument editor</extracomment>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Random</source>
+        <extracomment>Sample selection algorithm available in the instrument editor</extracomment>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>New component name</source>
+        <translation type="unfinished"></translation>
+    </message>
+</context>
+<context>
+    <name>ComponentsEditor</name>
+    <message>
+        <source>Component name</source>
+        <translation type="unfinished"></translation>
+    </message>
+</context>
+<context>
     <name>Director</name>
     <message>
         <source>Director</source>
@@ -1949,14 +1998,6 @@ Overwrite the existing file?</source>
 <context>
     <name>InstrumentEditor</name>
     <message>
-        <source>Show instrument properties</source>
-        <translation>Mostrar propietats de l&apos;instrument</translation>
-    </message>
-    <message>
-        <source>Show layers properties</source>
-        <translation>Mostrar propietats de les capes</translation>
-    </message>
-    <message>
         <source>Random pitch factor</source>
         <translation>Factor aleatorio de tono </translation>
     </message>
@@ -1969,20 +2010,8 @@ Overwrite the existing file?</source>
         <translation>Ressonància en el filtre</translation>
     </message>
     <message>
-        <source>Layer gain</source>
-        <translation>Guany de capa</translation>
-    </message>
-    <message>
         <source>Instrument gain</source>
         <translation>Guany de l&apos;instrument</translation>
-    </message>
-    <message>
-        <source>Layer pitch (Coarse)</source>
-        <translation>Ajust tò de capa (Fort)</translation>
-    </message>
-    <message>
-        <source>Layer pitch (Fine)</source>
-        <translation>Ajust tò de capa (Suau)</translation>
     </message>
     <message>
         <source>Midi out channel</source>
@@ -1997,14 +2026,6 @@ Overwrite the existing file?</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <source>Component volume</source>
-        <translation type="unfinished"></translation>
-    </message>
-    <message>
-        <source>Component name</source>
-        <translation type="unfinished"></translation>
-    </message>
-    <message>
         <source>Pitch offset (Coarse)</source>
         <translation type="unfinished"></translation>
     </message>
@@ -2014,10 +2035,6 @@ Overwrite the existing file?</source>
     </message>
     <message>
         <source>Stop the current playing instrument-note before trigger the next note sample</source>
-        <translation type="unfinished"></translation>
-    </message>
-    <message>
-        <source>Select selection algorithm</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
@@ -2044,23 +2061,15 @@ Value</source>
 Value</source>
         <translation type="unfinished"></translation>
     </message>
+</context>
+<context>
+    <name>InstrumentEditorPanel</name>
     <message>
-        <source>First in Velocity</source>
-        <extracomment>Sample selection algorithm available in the instrument editor</extracomment>
-        <translation type="unfinished"></translation>
+        <source>Show instrument properties</source>
+        <translation>Mostrar propietats de l&apos;instrument</translation>
     </message>
     <message>
-        <source>Round Robin</source>
-        <extracomment>Sample selection algorithm available in the instrument editor</extracomment>
-        <translation type="unfinished"></translation>
-    </message>
-    <message>
-        <source>Random</source>
-        <extracomment>Sample selection algorithm available in the instrument editor</extracomment>
-        <translation type="unfinished"></translation>
-    </message>
-    <message>
-        <source>New component name</source>
+        <source>Show components</source>
         <translation type="unfinished"></translation>
     </message>
 </context>
