+++ data/i18n/hydrogen_de.ts
@@ -1391,6 +1391,55 @@ Please set your system&apos;s locale to UTF-8!</source>
     </message>
 </context>
 <context>
+    <name>ComponentView</name>
+    <message>
+        <source>Layer gain</source>
+        <translation>Verstärkung (gain) der Ebene</translation>
+    </message>
+    <message>
+        <source>Component volume</source>
+        <translation>Lautstärke der Komponente</translation>
+    </message>
+    <message>
+        <source>Layer pitch (Coarse)</source>
+        <translation>Tonhöhe des Layers (grob)</translation>
+    </message>
+    <message>
+        <source>Layer pitch (Fine)</source>
+        <translation>Tonhöhe des Layers (fein)</translation>
+    </message>
+    <message>
+        <source>Select selection algorithm</source>
+        <translation>Wähle den Algorithmus zur Auswahl von Samples aus</translation>
+    </message>
+    <message>
+        <source>First in Velocity</source>
+        <extracomment>Sample selection algorithm available in the instrument editor</extracomment>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Round Robin</source>
+        <extracomment>Sample selection algorithm available in the instrument editor</extracomment>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Random</source>
+        <extracomment>Sample selection algorithm available in the instrument editor</extracomment>
+        <translation>Zufall</translation>
+    </message>
+    <message>
+        <source>New component name</source>
+        <translation>Neuer Komponentenname</translation>
+    </message>
+</context>
+<context>
+    <name>ComponentsEditor</name>
+    <message>
+        <source>Component name</source>
+        <translation>Komponentenname</translation>
+    </message>
+</context>
+<context>
     <name>Director</name>
     <message>
         <source>Director</source>
@@ -1951,14 +2000,6 @@ Overwrite the existing file?</source>
 <context>
     <name>InstrumentEditor</name>
     <message>
-        <source>Show instrument properties</source>
-        <translation>Eigenschaften des Instruments anzeigen</translation>
-    </message>
-    <message>
-        <source>Show layers properties</source>
-        <translation>Eigenschaften der Ebene anzeigen</translation>
-    </message>
-    <message>
         <source>Random pitch factor</source>
         <translation>Grad der zufälligen Tonhöhe</translation>
     </message>
@@ -1971,24 +2012,8 @@ Overwrite the existing file?</source>
         <translation>Resonanz des Filters</translation>
     </message>
     <message>
-        <source>Layer gain</source>
-        <translation>Verstärkung (gain) der Ebene</translation>
-    </message>
-    <message>
         <source>Instrument gain</source>
         <translation>Verstärkung (gain) des Instruments</translation>
-    </message>
-    <message>
-        <source>Layer pitch (Coarse)</source>
-        <translation>Tonhöhe des Layers (grob)</translation>
-    </message>
-    <message>
-        <source>Layer pitch (Fine)</source>
-        <translation>Tonhöhe des Layers (fein)</translation>
-    </message>
-    <message>
-        <source>Component name</source>
-        <translation>Komponentenname</translation>
     </message>
     <message>
         <source>Midi out channel</source>
@@ -1997,10 +2022,6 @@ Overwrite the existing file?</source>
     <message>
         <source>Midi out note</source>
         <translation>Note der Midi-Ausgabe</translation>
-    </message>
-    <message>
-        <source>Component volume</source>
-        <translation>Lautstärke der Komponente</translation>
     </message>
     <message>
         <source>Don&apos;t change the layers&apos; gain based on velocity</source>
@@ -2017,10 +2038,6 @@ Overwrite the existing file?</source>
     <message>
         <source>Stop the current playing instrument-note before trigger the next note sample</source>
         <translation>Stoppe die aktuell abgespielte Note für dieses Instrument bevor nächste Note abgespielt wird</translation>
-    </message>
-    <message>
-        <source>Select selection algorithm</source>
-        <translation>Wähle den Algorithmus zur Auswahl von Samples aus</translation>
     </message>
     <message>
         <source>Length of Attack phase.
@@ -2054,24 +2071,16 @@ Value</source>
 
 Wert</translation>
     </message>
+</context>
+<context>
+    <name>InstrumentEditorPanel</name>
     <message>
-        <source>First in Velocity</source>
-        <extracomment>Sample selection algorithm available in the instrument editor</extracomment>
+        <source>Show instrument properties</source>
+        <translation>Eigenschaften des Instruments anzeigen</translation>
+    </message>
+    <message>
+        <source>Show components</source>
         <translation type="unfinished"></translation>
-    </message>
-    <message>
-        <source>Round Robin</source>
-        <extracomment>Sample selection algorithm available in the instrument editor</extracomment>
-        <translation type="unfinished"></translation>
-    </message>
-    <message>
-        <source>Random</source>
-        <extracomment>Sample selection algorithm available in the instrument editor</extracomment>
-        <translation>Zufall</translation>
-    </message>
-    <message>
-        <source>New component name</source>
-        <translation>Neuer Komponentenname</translation>
     </message>
 </context>
 <context>
