+++ data/i18n/hydrogen_el.ts
@@ -1388,6 +1388,55 @@ Please set your system&apos;s locale to UTF-8!</source>
     </message>
 </context>
 <context>
+    <name>ComponentView</name>
+    <message>
+        <source>Layer gain</source>
+        <translation>Ένταση στρώματος</translation>
+    </message>
+    <message>
+        <source>Component volume</source>
+        <translation>Ένταση του συστατικού</translation>
+    </message>
+    <message>
+        <source>Layer pitch (Coarse)</source>
+        <translation>Τόνος στρώματος (Τραχύς)</translation>
+    </message>
+    <message>
+        <source>Layer pitch (Fine)</source>
+        <translation>Τόνος του στρώματος (Φίνος)</translation>
+    </message>
+    <message>
+        <source>Select selection algorithm</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>First in Velocity</source>
+        <extracomment>Sample selection algorithm available in the instrument editor</extracomment>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Round Robin</source>
+        <extracomment>Sample selection algorithm available in the instrument editor</extracomment>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Random</source>
+        <extracomment>Sample selection algorithm available in the instrument editor</extracomment>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>New component name</source>
+        <translation>Όνομα του νέου συστατικού</translation>
+    </message>
+</context>
+<context>
+    <name>ComponentsEditor</name>
+    <message>
+        <source>Component name</source>
+        <translation>Όνομα του συστατικού</translation>
+    </message>
+</context>
+<context>
     <name>Director</name>
     <message>
         <source>Director</source>
@@ -1948,14 +1997,6 @@ Overwrite the existing file?</source>
 <context>
     <name>InstrumentEditor</name>
     <message>
-        <source>Show instrument properties</source>
-        <translation>Προβολή των ιδιοτήτων του οργάνου</translation>
-    </message>
-    <message>
-        <source>Show layers properties</source>
-        <translation>Προβολή των ιδιοτήτων των στρωμάτων</translation>
-    </message>
-    <message>
         <source>Random pitch factor</source>
         <translation>Παράγοντας τυχαίας τονικότητας</translation>
     </message>
@@ -1972,32 +2013,12 @@ Overwrite the existing file?</source>
         <translation>Ένταση οργάνου</translation>
     </message>
     <message>
-        <source>Layer gain</source>
-        <translation>Ένταση στρώματος</translation>
-    </message>
-    <message>
-        <source>Layer pitch (Coarse)</source>
-        <translation>Τόνος στρώματος (Τραχύς)</translation>
-    </message>
-    <message>
-        <source>Layer pitch (Fine)</source>
-        <translation>Τόνος του στρώματος (Φίνος)</translation>
-    </message>
-    <message>
-        <source>Component name</source>
-        <translation>Όνομα του συστατικού</translation>
-    </message>
-    <message>
         <source>Midi out channel</source>
         <translation>Κανάλι εξόδου ΜΙΝΤΙ</translation>
     </message>
     <message>
         <source>Midi out note</source>
         <translation>Στίγμα εξόδου ΜΙΝΤΙ</translation>
-    </message>
-    <message>
-        <source>Component volume</source>
-        <translation>Ένταση του συστατικού</translation>
     </message>
     <message>
         <source>Don&apos;t change the layers&apos; gain based on velocity</source>
@@ -2013,10 +2034,6 @@ Overwrite the existing file?</source>
     </message>
     <message>
         <source>Stop the current playing instrument-note before trigger the next note sample</source>
-        <translation type="unfinished"></translation>
-    </message>
-    <message>
-        <source>Select selection algorithm</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
@@ -2043,24 +2060,16 @@ Value</source>
 Value</source>
         <translation type="unfinished"></translation>
     </message>
+</context>
+<context>
+    <name>InstrumentEditorPanel</name>
     <message>
-        <source>First in Velocity</source>
-        <extracomment>Sample selection algorithm available in the instrument editor</extracomment>
-        <translation type="unfinished"></translation>
+        <source>Show instrument properties</source>
+        <translation>Προβολή των ιδιοτήτων του οργάνου</translation>
     </message>
     <message>
-        <source>Round Robin</source>
-        <extracomment>Sample selection algorithm available in the instrument editor</extracomment>
+        <source>Show components</source>
         <translation type="unfinished"></translation>
-    </message>
-    <message>
-        <source>Random</source>
-        <extracomment>Sample selection algorithm available in the instrument editor</extracomment>
-        <translation type="unfinished"></translation>
-    </message>
-    <message>
-        <source>New component name</source>
-        <translation>Όνομα του νέου συστατικού</translation>
     </message>
 </context>
 <context>
