+++ data/i18n/hydrogen_en.ts
@@ -1388,6 +1388,55 @@ Please set your system&apos;s locale to UTF-8!</source>
     </message>
 </context>
 <context>
+    <name>ComponentView</name>
+    <message>
+        <source>Layer gain</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Component volume</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Layer pitch (Coarse)</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Layer pitch (Fine)</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Select selection algorithm</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>First in Velocity</source>
+        <extracomment>Sample selection algorithm available in the instrument editor</extracomment>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Round Robin</source>
+        <extracomment>Sample selection algorithm available in the instrument editor</extracomment>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Random</source>
+        <extracomment>Sample selection algorithm available in the instrument editor</extracomment>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>New component name</source>
+        <translation type="unfinished"></translation>
+    </message>
+</context>
+<context>
+    <name>ComponentsEditor</name>
+    <message>
+        <source>Component name</source>
+        <translation type="unfinished"></translation>
+    </message>
+</context>
+<context>
     <name>Director</name>
     <message>
         <source>Director</source>
@@ -1948,14 +1997,6 @@ Overwrite the existing file?</source>
 <context>
     <name>InstrumentEditor</name>
     <message>
-        <source>Show instrument properties</source>
-        <translation></translation>
-    </message>
-    <message>
-        <source>Show layers properties</source>
-        <translation></translation>
-    </message>
-    <message>
         <source>Midi out channel</source>
         <translation></translation>
     </message>
@@ -1984,26 +2025,6 @@ Overwrite the existing file?</source>
         <translation></translation>
     </message>
     <message>
-        <source>Layer gain</source>
-        <translation></translation>
-    </message>
-    <message>
-        <source>Component volume</source>
-        <translation></translation>
-    </message>
-    <message>
-        <source>Layer pitch (Coarse)</source>
-        <translation></translation>
-    </message>
-    <message>
-        <source>Layer pitch (Fine)</source>
-        <translation></translation>
-    </message>
-    <message>
-        <source>Component name</source>
-        <translation></translation>
-    </message>
-    <message>
         <source>Pitch offset (Coarse)</source>
         <translation></translation>
     </message>
@@ -2014,10 +2035,6 @@ Overwrite the existing file?</source>
     <message>
         <source>Stop the current playing instrument-note before trigger the next note sample</source>
         <translation>Stop the current playing instrument-note before triggering the next note sample</translation>
-    </message>
-    <message>
-        <source>Select selection algorithm</source>
-        <translation></translation>
     </message>
     <message>
         <source>Length of Attack phase.
@@ -2043,24 +2060,16 @@ Value</source>
 Value</source>
         <translation></translation>
     </message>
+</context>
+<context>
+    <name>InstrumentEditorPanel</name>
     <message>
-        <source>First in Velocity</source>
-        <extracomment>Sample selection algorithm available in the instrument editor</extracomment>
-        <translation></translation>
+        <source>Show instrument properties</source>
+        <translation type="unfinished"></translation>
     </message>
     <message>
-        <source>Round Robin</source>
-        <extracomment>Sample selection algorithm available in the instrument editor</extracomment>
-        <translation></translation>
-    </message>
-    <message>
-        <source>Random</source>
-        <extracomment>Sample selection algorithm available in the instrument editor</extracomment>
-        <translation></translation>
-    </message>
-    <message>
-        <source>New component name</source>
-        <translation></translation>
+        <source>Show components</source>
+        <translation type="unfinished"></translation>
     </message>
 </context>
 <context>
