+++ data/i18n/hydrogen_es.ts
@@ -1392,6 +1392,55 @@ Please set your system&apos;s locale to UTF-8!</source>
     </message>
 </context>
 <context>
+    <name>ComponentView</name>
+    <message>
+        <source>Layer gain</source>
+        <translation>Ganancia de capa</translation>
+    </message>
+    <message>
+        <source>Component volume</source>
+        <translation>Volumen componente</translation>
+    </message>
+    <message>
+        <source>Layer pitch (Coarse)</source>
+        <translation>Afinación de capa (gruesa)</translation>
+    </message>
+    <message>
+        <source>Layer pitch (Fine)</source>
+        <translation>Afinación de capa (fina)</translation>
+    </message>
+    <message>
+        <source>Select selection algorithm</source>
+        <translation>Seleccionar algoritmo de selección</translation>
+    </message>
+    <message>
+        <source>First in Velocity</source>
+        <extracomment>Sample selection algorithm available in the instrument editor</extracomment>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Round Robin</source>
+        <extracomment>Sample selection algorithm available in the instrument editor</extracomment>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Random</source>
+        <extracomment>Sample selection algorithm available in the instrument editor</extracomment>
+        <translation>Azar</translation>
+    </message>
+    <message>
+        <source>New component name</source>
+        <translation>Nombre nuevo componente</translation>
+    </message>
+</context>
+<context>
+    <name>ComponentsEditor</name>
+    <message>
+        <source>Component name</source>
+        <translation>Nombre componente</translation>
+    </message>
+</context>
+<context>
     <name>Director</name>
     <message>
         <source>Director</source>
@@ -1954,14 +2003,6 @@ Overwrite the existing file?</source>
 <context>
     <name>InstrumentEditor</name>
     <message>
-        <source>Show instrument properties</source>
-        <translation>Mostrar propiedades de instrumento</translation>
-    </message>
-    <message>
-        <source>Show layers properties</source>
-        <translation>Mostrar propiedades de capas</translation>
-    </message>
-    <message>
         <source>Random pitch factor</source>
         <translation>Factor aleatorio de tono</translation>
     </message>
@@ -1974,24 +2015,8 @@ Overwrite the existing file?</source>
         <translation>Filtro de resonancia</translation>
     </message>
     <message>
-        <source>Layer gain</source>
-        <translation>Ganancia de capa</translation>
-    </message>
-    <message>
         <source>Instrument gain</source>
         <translation>Ganancia del instrumento</translation>
-    </message>
-    <message>
-        <source>Layer pitch (Coarse)</source>
-        <translation>Afinación de capa (gruesa)</translation>
-    </message>
-    <message>
-        <source>Layer pitch (Fine)</source>
-        <translation>Afinación de capa (fina)</translation>
-    </message>
-    <message>
-        <source>Component name</source>
-        <translation>Nombre componente</translation>
     </message>
     <message>
         <source>Midi out channel</source>
@@ -2000,10 +2025,6 @@ Overwrite the existing file?</source>
     <message>
         <source>Midi out note</source>
         <translation>Nota salida Midi</translation>
-    </message>
-    <message>
-        <source>Component volume</source>
-        <translation>Volumen componente</translation>
     </message>
     <message>
         <source>Don&apos;t change the layers&apos; gain based on velocity</source>
@@ -2020,10 +2041,6 @@ Overwrite the existing file?</source>
     <message>
         <source>Stop the current playing instrument-note before trigger the next note sample</source>
         <translation>Detén la nota de instrumento actualmente reproduciéndose antes de disparar el sample de la siguiente nota</translation>
-    </message>
-    <message>
-        <source>Select selection algorithm</source>
-        <translation>Seleccionar algoritmo de selección</translation>
     </message>
     <message>
         <source>Length of Attack phase.
@@ -2057,24 +2074,16 @@ Value</source>
 
 Valor</translation>
     </message>
+</context>
+<context>
+    <name>InstrumentEditorPanel</name>
     <message>
-        <source>First in Velocity</source>
-        <extracomment>Sample selection algorithm available in the instrument editor</extracomment>
+        <source>Show instrument properties</source>
+        <translation>Mostrar propiedades de instrumento</translation>
+    </message>
+    <message>
+        <source>Show components</source>
         <translation type="unfinished"></translation>
-    </message>
-    <message>
-        <source>Round Robin</source>
-        <extracomment>Sample selection algorithm available in the instrument editor</extracomment>
-        <translation type="unfinished"></translation>
-    </message>
-    <message>
-        <source>Random</source>
-        <extracomment>Sample selection algorithm available in the instrument editor</extracomment>
-        <translation>Azar</translation>
-    </message>
-    <message>
-        <source>New component name</source>
-        <translation>Nombre nuevo componente</translation>
     </message>
 </context>
 <context>
