+++ data/i18n/hydrogen_hr.ts
@@ -1388,6 +1388,55 @@ Please set your system&apos;s locale to UTF-8!</source>
     </message>
 </context>
 <context>
+    <name>ComponentView</name>
+    <message>
+        <source>Layer gain</source>
+        <translation>Gain razine</translation>
+    </message>
+    <message>
+        <source>Component volume</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Layer pitch (Coarse)</source>
+        <translation>Layer pitch (grezzo)</translation>
+    </message>
+    <message>
+        <source>Layer pitch (Fine)</source>
+        <translation>Layer pitch (Fine)</translation>
+    </message>
+    <message>
+        <source>Select selection algorithm</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>First in Velocity</source>
+        <extracomment>Sample selection algorithm available in the instrument editor</extracomment>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Round Robin</source>
+        <extracomment>Sample selection algorithm available in the instrument editor</extracomment>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Random</source>
+        <extracomment>Sample selection algorithm available in the instrument editor</extracomment>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>New component name</source>
+        <translation type="unfinished"></translation>
+    </message>
+</context>
+<context>
+    <name>ComponentsEditor</name>
+    <message>
+        <source>Component name</source>
+        <translation type="unfinished"></translation>
+    </message>
+</context>
+<context>
     <name>Director</name>
     <message>
         <source>Director</source>
@@ -1948,18 +1997,6 @@ Overwrite the existing file?</source>
 <context>
     <name>InstrumentEditor</name>
     <message>
-        <source>Show instrument properties</source>
-        <translation>Prikaži karakteristike instrumenta</translation>
-    </message>
-    <message>
-        <source>Show layers properties</source>
-        <translation>Prikaži karakteristike razine</translation>
-    </message>
-    <message>
-        <source>Layer gain</source>
-        <translation>Gain razine</translation>
-    </message>
-    <message>
         <source>Instrument gain</source>
         <translation>Gain instrumenta</translation>
     </message>
@@ -1976,27 +2013,11 @@ Overwrite the existing file?</source>
         <translation>Filter rezonancija</translation>
     </message>
     <message>
-        <source>Layer pitch (Coarse)</source>
-        <translation>Layer pitch (grezzo)</translation>
-    </message>
-    <message>
-        <source>Layer pitch (Fine)</source>
-        <translation>Layer pitch (Fine)</translation>
-    </message>
-    <message>
-        <source>Component name</source>
-        <translation type="unfinished"></translation>
-    </message>
-    <message>
         <source>Midi out channel</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
         <source>Midi out note</source>
-        <translation type="unfinished"></translation>
-    </message>
-    <message>
-        <source>Component volume</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
@@ -2013,10 +2034,6 @@ Overwrite the existing file?</source>
     </message>
     <message>
         <source>Stop the current playing instrument-note before trigger the next note sample</source>
-        <translation type="unfinished"></translation>
-    </message>
-    <message>
-        <source>Select selection algorithm</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
@@ -2043,23 +2060,15 @@ Value</source>
 Value</source>
         <translation type="unfinished"></translation>
     </message>
+</context>
+<context>
+    <name>InstrumentEditorPanel</name>
     <message>
-        <source>First in Velocity</source>
-        <extracomment>Sample selection algorithm available in the instrument editor</extracomment>
-        <translation type="unfinished"></translation>
+        <source>Show instrument properties</source>
+        <translation>Prikaži karakteristike instrumenta</translation>
     </message>
     <message>
-        <source>Round Robin</source>
-        <extracomment>Sample selection algorithm available in the instrument editor</extracomment>
-        <translation type="unfinished"></translation>
-    </message>
-    <message>
-        <source>Random</source>
-        <extracomment>Sample selection algorithm available in the instrument editor</extracomment>
-        <translation type="unfinished"></translation>
-    </message>
-    <message>
-        <source>New component name</source>
+        <source>Show components</source>
         <translation type="unfinished"></translation>
     </message>
 </context>
