+++ data/i18n/hydrogen_pt_BR.ts
@@ -1391,6 +1391,55 @@ Please set your system&apos;s locale to UTF-8!</source>
     </message>
 </context>
 <context>
+    <name>ComponentView</name>
+    <message>
+        <source>Layer gain</source>
+        <translation>Ganho na camada</translation>
+    </message>
+    <message>
+        <source>Component volume</source>
+        <translation>Volume do componente</translation>
+    </message>
+    <message>
+        <source>Layer pitch (Coarse)</source>
+        <translation>Tom da camada (Grosseiro)</translation>
+    </message>
+    <message>
+        <source>Layer pitch (Fine)</source>
+        <translation>Tom da camada (Fino)</translation>
+    </message>
+    <message>
+        <source>Select selection algorithm</source>
+        <translation>Selecionar algoritmo de seleção</translation>
+    </message>
+    <message>
+        <source>First in Velocity</source>
+        <extracomment>Sample selection algorithm available in the instrument editor</extracomment>
+        <translation>Primeiro em Velocidade</translation>
+    </message>
+    <message>
+        <source>Round Robin</source>
+        <extracomment>Sample selection algorithm available in the instrument editor</extracomment>
+        <translation>Round Robin</translation>
+    </message>
+    <message>
+        <source>Random</source>
+        <extracomment>Sample selection algorithm available in the instrument editor</extracomment>
+        <translation>Aleatório</translation>
+    </message>
+    <message>
+        <source>New component name</source>
+        <translation>Novo nome do componente</translation>
+    </message>
+</context>
+<context>
+    <name>ComponentsEditor</name>
+    <message>
+        <source>Component name</source>
+        <translation>Nome do componente</translation>
+    </message>
+</context>
+<context>
     <name>Director</name>
     <message>
         <source>Director</source>
@@ -1953,14 +2002,6 @@ Sobrescrever o arquivo existente?</translation>
 <context>
     <name>InstrumentEditor</name>
     <message>
-        <source>Show instrument properties</source>
-        <translation>Exibir propriedades do instrumento</translation>
-    </message>
-    <message>
-        <source>Show layers properties</source>
-        <translation>Exibir propriedades das camadas</translation>
-    </message>
-    <message>
         <source>Random pitch factor</source>
         <translation>Tom aleatório</translation>
     </message>
@@ -1973,24 +2014,8 @@ Sobrescrever o arquivo existente?</translation>
         <translation>Filtro de ressonância</translation>
     </message>
     <message>
-        <source>Layer gain</source>
-        <translation>Ganho na camada</translation>
-    </message>
-    <message>
         <source>Instrument gain</source>
         <translation>Ganho do instrumento</translation>
-    </message>
-    <message>
-        <source>Layer pitch (Coarse)</source>
-        <translation>Tom da camada (Grosseiro)</translation>
-    </message>
-    <message>
-        <source>Layer pitch (Fine)</source>
-        <translation>Tom da camada (Fino)</translation>
-    </message>
-    <message>
-        <source>Component name</source>
-        <translation>Nome do componente</translation>
     </message>
     <message>
         <source>Midi out channel</source>
@@ -1999,10 +2024,6 @@ Sobrescrever o arquivo existente?</translation>
     <message>
         <source>Midi out note</source>
         <translation>Nota de saída MIDI</translation>
-    </message>
-    <message>
-        <source>Component volume</source>
-        <translation>Volume do componente</translation>
     </message>
     <message>
         <source>Don&apos;t change the layers&apos; gain based on velocity</source>
@@ -2019,10 +2040,6 @@ Sobrescrever o arquivo existente?</translation>
     <message>
         <source>Stop the current playing instrument-note before trigger the next note sample</source>
         <translation>Pare de tocar a nota do instrumento atual antes de iniciar a próxima nota de amostra</translation>
-    </message>
-    <message>
-        <source>Select selection algorithm</source>
-        <translation>Selecionar algoritmo de seleção</translation>
     </message>
     <message>
         <source>Length of Attack phase.
@@ -2056,24 +2073,16 @@ Value</source>
 
 Valor</translation>
     </message>
+</context>
+<context>
+    <name>InstrumentEditorPanel</name>
     <message>
-        <source>First in Velocity</source>
-        <extracomment>Sample selection algorithm available in the instrument editor</extracomment>
-        <translation>Primeiro em Velocidade</translation>
+        <source>Show instrument properties</source>
+        <translation>Exibir propriedades do instrumento</translation>
     </message>
     <message>
-        <source>Round Robin</source>
-        <extracomment>Sample selection algorithm available in the instrument editor</extracomment>
-        <translation>Round Robin</translation>
-    </message>
-    <message>
-        <source>Random</source>
-        <extracomment>Sample selection algorithm available in the instrument editor</extracomment>
-        <translation>Aleatório</translation>
-    </message>
-    <message>
-        <source>New component name</source>
-        <translation>Novo nome do componente</translation>
+        <source>Show components</source>
+        <translation type="unfinished"></translation>
     </message>
 </context>
 <context>
