+++ data/i18n/hydrogen_ru.ts
@@ -1388,6 +1388,55 @@ Please set your system&apos;s locale to UTF-8!</source>
     </message>
 </context>
 <context>
+    <name>ComponentView</name>
+    <message>
+        <source>Layer gain</source>
+        <translation>Усиление слоя</translation>
+    </message>
+    <message>
+        <source>Component volume</source>
+        <translation>Громкость компонента</translation>
+    </message>
+    <message>
+        <source>Layer pitch (Coarse)</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Layer pitch (Fine)</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Select selection algorithm</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>First in Velocity</source>
+        <extracomment>Sample selection algorithm available in the instrument editor</extracomment>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Round Robin</source>
+        <extracomment>Sample selection algorithm available in the instrument editor</extracomment>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Random</source>
+        <extracomment>Sample selection algorithm available in the instrument editor</extracomment>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>New component name</source>
+        <translation>Новое название компонента</translation>
+    </message>
+</context>
+<context>
+    <name>ComponentsEditor</name>
+    <message>
+        <source>Component name</source>
+        <translation>Название компонента</translation>
+    </message>
+</context>
+<context>
     <name>Director</name>
     <message>
         <source>Director</source>
@@ -1950,24 +1999,12 @@ Overwrite the existing file?</source>
 <context>
     <name>InstrumentEditor</name>
     <message>
-        <source>Show instrument properties</source>
-        <translation>Показать свойства инструмента</translation>
-    </message>
-    <message>
-        <source>Show layers properties</source>
-        <translation>Показать свойства слоя</translation>
-    </message>
-    <message>
         <source>Random pitch factor</source>
         <translation>Случайный коэфф. высоты тона</translation>
     </message>
     <message>
         <source>Filter resonance</source>
         <translation>Резонансный фильтр</translation>
-    </message>
-    <message>
-        <source>Layer gain</source>
-        <translation>Усиление слоя</translation>
     </message>
     <message>
         <source>Instrument gain</source>
@@ -1978,28 +2015,12 @@ Overwrite the existing file?</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <source>Layer pitch (Coarse)</source>
-        <translation type="unfinished"></translation>
-    </message>
-    <message>
-        <source>Layer pitch (Fine)</source>
-        <translation type="unfinished"></translation>
-    </message>
-    <message>
-        <source>Component name</source>
-        <translation>Название компонента</translation>
-    </message>
-    <message>
         <source>Midi out channel</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
         <source>Midi out note</source>
         <translation type="unfinished"></translation>
-    </message>
-    <message>
-        <source>Component volume</source>
-        <translation>Громкость компонента</translation>
     </message>
     <message>
         <source>Don&apos;t change the layers&apos; gain based on velocity</source>
@@ -2015,10 +2036,6 @@ Overwrite the existing file?</source>
     </message>
     <message>
         <source>Stop the current playing instrument-note before trigger the next note sample</source>
-        <translation type="unfinished"></translation>
-    </message>
-    <message>
-        <source>Select selection algorithm</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
@@ -2045,24 +2062,16 @@ Value</source>
 Value</source>
         <translation type="unfinished"></translation>
     </message>
+</context>
+<context>
+    <name>InstrumentEditorPanel</name>
     <message>
-        <source>First in Velocity</source>
-        <extracomment>Sample selection algorithm available in the instrument editor</extracomment>
-        <translation type="unfinished"></translation>
+        <source>Show instrument properties</source>
+        <translation>Показать свойства инструмента</translation>
     </message>
     <message>
-        <source>Round Robin</source>
-        <extracomment>Sample selection algorithm available in the instrument editor</extracomment>
+        <source>Show components</source>
         <translation type="unfinished"></translation>
-    </message>
-    <message>
-        <source>Random</source>
-        <extracomment>Sample selection algorithm available in the instrument editor</extracomment>
-        <translation type="unfinished"></translation>
-    </message>
-    <message>
-        <source>New component name</source>
-        <translation>Новое название компонента</translation>
     </message>
 </context>
 <context>
