+++ data/i18n/hydrogen_uk.ts
@@ -1388,6 +1388,55 @@ Please set your system&apos;s locale to UTF-8!</source>
     </message>
 </context>
 <context>
+    <name>ComponentView</name>
+    <message>
+        <source>Layer gain</source>
+        <translation>Підсилення шару</translation>
+    </message>
+    <message>
+        <source>Component volume</source>
+        <translation>Гучність компонента</translation>
+    </message>
+    <message>
+        <source>Layer pitch (Coarse)</source>
+        <translation>Тон шару (грубо)</translation>
+    </message>
+    <message>
+        <source>Layer pitch (Fine)</source>
+        <translation>Тон шару (точно)</translation>
+    </message>
+    <message>
+        <source>Select selection algorithm</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>First in Velocity</source>
+        <extracomment>Sample selection algorithm available in the instrument editor</extracomment>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Round Robin</source>
+        <extracomment>Sample selection algorithm available in the instrument editor</extracomment>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Random</source>
+        <extracomment>Sample selection algorithm available in the instrument editor</extracomment>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>New component name</source>
+        <translation>Назва нового компонента</translation>
+    </message>
+</context>
+<context>
+    <name>ComponentsEditor</name>
+    <message>
+        <source>Component name</source>
+        <translation>Назва компонента</translation>
+    </message>
+</context>
+<context>
     <name>Director</name>
     <message>
         <source>Director</source>
@@ -1950,24 +1999,12 @@ Overwrite the existing file?</source>
 <context>
     <name>InstrumentEditor</name>
     <message>
-        <source>Show instrument properties</source>
-        <translation>Показати властивості інструмента</translation>
-    </message>
-    <message>
-        <source>Show layers properties</source>
-        <translation>Показати властивості шару</translation>
-    </message>
-    <message>
         <source>Random pitch factor</source>
         <translation>Випадковий коеф. висоти тону</translation>
     </message>
     <message>
         <source>Filter resonance</source>
         <translation>Резонансний фільтр</translation>
-    </message>
-    <message>
-        <source>Layer gain</source>
-        <translation>Підсилення шару</translation>
     </message>
     <message>
         <source>Instrument gain</source>
@@ -1978,28 +2015,12 @@ Overwrite the existing file?</source>
         <translation>Фільтр відсікання</translation>
     </message>
     <message>
-        <source>Layer pitch (Coarse)</source>
-        <translation>Тон шару (грубо)</translation>
-    </message>
-    <message>
-        <source>Layer pitch (Fine)</source>
-        <translation>Тон шару (точно)</translation>
-    </message>
-    <message>
-        <source>Component name</source>
-        <translation>Назва компонента</translation>
-    </message>
-    <message>
         <source>Midi out channel</source>
         <translation>Вихідний канал MIDI</translation>
     </message>
     <message>
         <source>Midi out note</source>
         <translation>Вихідна нота MIDI</translation>
-    </message>
-    <message>
-        <source>Component volume</source>
-        <translation>Гучність компонента</translation>
     </message>
     <message>
         <source>Don&apos;t change the layers&apos; gain based on velocity</source>
@@ -2015,10 +2036,6 @@ Overwrite the existing file?</source>
     </message>
     <message>
         <source>Stop the current playing instrument-note before trigger the next note sample</source>
-        <translation type="unfinished"></translation>
-    </message>
-    <message>
-        <source>Select selection algorithm</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
@@ -2045,24 +2062,16 @@ Value</source>
 Value</source>
         <translation type="unfinished"></translation>
     </message>
+</context>
+<context>
+    <name>InstrumentEditorPanel</name>
     <message>
-        <source>First in Velocity</source>
-        <extracomment>Sample selection algorithm available in the instrument editor</extracomment>
-        <translation type="unfinished"></translation>
+        <source>Show instrument properties</source>
+        <translation>Показати властивості інструмента</translation>
     </message>
     <message>
-        <source>Round Robin</source>
-        <extracomment>Sample selection algorithm available in the instrument editor</extracomment>
+        <source>Show components</source>
         <translation type="unfinished"></translation>
-    </message>
-    <message>
-        <source>Random</source>
-        <extracomment>Sample selection algorithm available in the instrument editor</extracomment>
-        <translation type="unfinished"></translation>
-    </message>
-    <message>
-        <source>New component name</source>
-        <translation>Назва нового компонента</translation>
     </message>
 </context>
 <context>
