+++ src/gui/src/InstrumentEditor/InstrumentEditor.cpp
@@ -20,47 +20,29 @@
  *
  */
 
-#include <QtGui>
-#include <QtWidgets>
+#include "InstrumentEditor.h"
 
-
-#include <math.h>
-#include <assert.h>
-#include <vector>
-
-#include <core/Hydrogen.h>
-#include <core/Globals.h>
 #include <core/Basics/Adsr.h>
 #include <core/Basics/Drumkit.h>
+#include <core/Basics/Event.h>
 #include <core/Basics/Instrument.h>
 #include <core/Basics/InstrumentList.h>
-#include <core/Basics/InstrumentComponent.h>
-#include <core/Basics/InstrumentLayer.h>
-#include <core/Basics/Sample.h>
 #include <core/Basics/Song.h>
-#include <core/AudioEngine/AudioEngine.h>
-#include <core/EventQueue.h>
-using namespace H2Core;
+#include <core/Hydrogen.h>
 
-#include "../HydrogenApp.h"
+#include "InstrumentEditorPanel.h"
 #include "../CommonStrings.h"
-#include "../UndoActions.h"
-#include "../Widgets/Rotary.h"
-#include "../Widgets/WidgetWithInput.h"
-#include "../Widgets/ClickableLabel.h"
+#include "../HydrogenApp.h"
 #include "../Widgets/Button.h"
+#include "../Widgets/ClickableLabel.h"
 #include "../Widgets/LCDDisplay.h"
 #include "../Widgets/LCDSpinBox.h"
-#include "../Widgets/LCDCombo.h"
-#include "InstrumentEditor.h"
-#include "InstrumentEditorPanel.h"
-#include "WaveDisplay.h"
-#include "LayerPreview.h"
-#include "AudioFileBrowser/AudioFileBrowser.h"
+#include "../Widgets/Rotary.h"
 
-InstrumentEditor::InstrumentEditor( QWidget* pParent,
-									InstrumentEditorPanel* pPanel )
-	: QWidget( pParent )
+using namespace H2Core;
+
+InstrumentEditor::InstrumentEditor( InstrumentEditorPanel* pPanel )
+	: QWidget( pPanel )
 	, m_pInstrumentEditorPanel( pPanel )
 	, m_fPreviousMidiOutChannel( -1.0 )
 {
@@ -69,28 +51,8 @@ InstrumentEditor::InstrumentEditor( QWidget* pParent,
 	auto pCommonStrings = HydrogenApp::get_instance()->getCommonStrings();
 	
 	// Instrument properties top
-	m_pInstrumentPropTop = new PixmapWidget( this );
-	m_pInstrumentPropTop->setPixmap( "/instrumentEditor/instrumentTab_top.png" );
-
-	m_pShowInstrumentBtn = new Button( m_pInstrumentPropTop, QSize( 141, 22 ), Button::Type::Toggle, "",
-									   pCommonStrings->getGeneralButton(), false, QSize(),
-									   tr( "Show instrument properties" ) );
-	m_pShowInstrumentBtn->move( 4, 4 );
-	
-	connect( m_pShowInstrumentBtn, &QPushButton::clicked,
-			 [=]() { showLayers( false ); } );
-
-	m_pShowLayersBtn = new Button( m_pInstrumentPropTop, QSize( 140, 22 ), Button::Type::Toggle, "",
-								   pCommonStrings->getLayersButton(), false, QSize(),
-								   tr( "Show layers properties" ) );
-	m_pShowLayersBtn->move( 145, 4 );
-	
-	connect( m_pShowLayersBtn, &QPushButton::clicked,
-			 [=]() { showLayers( true ); } );
-
-	// Instrument properties
 	m_pInstrumentProp = new PixmapWidget( this );
-	m_pInstrumentProp->move(0, 31);
+	m_pInstrumentProp->move( 0, 0 );
 	m_pInstrumentProp->setPixmap( "/instrumentEditor/instrumentTab.png" );
 
 	m_pNameLbl = new ClickableLabel( m_pInstrumentProp, QSize( 279, 27 ), "",
@@ -146,11 +108,10 @@ InstrumentEditor::InstrumentEditor( QWidget* pParent,
 									  pCommonStrings->getPitchLabel() );
 	m_pPitchLbl->move( 25, 235 );
 	
-	const float fFinePitch = 0.5;
-	m_pPitchCoarseRotary = new Rotary( m_pInstrumentProp, Rotary::Type::Center,
-									   tr( "Pitch offset (Coarse)" ), true,
-									   Instrument::fPitchMin + fFinePitch,
-									   Instrument::fPitchMax - fFinePitch );
+	m_pPitchCoarseRotary = new Rotary(
+		m_pInstrumentProp, Rotary::Type::Center, tr( "Pitch offset (Coarse)" ),
+		true, Instrument::fPitchMin + InstrumentEditorPanel::nPitchFineControl,
+		Instrument::fPitchMax - InstrumentEditorPanel::nPitchFineControl );
 	m_pPitchCoarseRotary->move( 84, 210 );
 	connect( m_pPitchCoarseRotary, &Rotary::valueChanged, [&]() {
 		//round fVal, since Coarse is the integer number of half steps
@@ -163,9 +124,10 @@ InstrumentEditor::InstrumentEditor( QWidget* pParent,
 											pCommonStrings->getPitchCoarseLabel() );
 	m_pPitchCoarseLbl->move( 82, 235 );
 
-	m_pPitchFineRotary = new Rotary( m_pInstrumentProp, Rotary::Type::Center,
-									 tr( "Pitch offset (Fine)" ), false,
-									 -fFinePitch, fFinePitch );
+	m_pPitchFineRotary = new Rotary(
+		m_pInstrumentProp, Rotary::Type::Center, tr( "Pitch offset (Fine)" ),
+		false, -InstrumentEditorPanel::nPitchFineControl,
+		InstrumentEditorPanel::nPitchFineControl );
 	//it will have resolution of 100 steps between Min and Max => quantum delta = 0.01
 	m_pPitchFineRotary->move( 138, 210 );
 	connect( m_pPitchFineRotary, &Rotary::valueChanged, [&]() {
@@ -373,208 +335,19 @@ InstrumentEditor::InstrumentEditor( QWidget* pParent,
 	m_pHihatMaxRangeLbl = new ClickableLabel( m_pInstrumentProp, QSize( 61, 10 ),
 											  pCommonStrings->getHihatMaxRangeLabel() );
 	m_pHihatMaxRangeLbl->move( 201, 327 );
-	// ~ Instrument properties
 
-	// LAYER properties
-	m_pLayerProp = new PixmapWidget( this );
-	m_pLayerProp->setObjectName( "LayerProperties" );
-	m_pLayerProp->move( 0, 31 );
-	m_pLayerProp->hide();
-	m_pLayerProp->setPixmap( "/instrumentEditor/layerTabsupernew.png" );
-
-	// Component
-	m_pCompoNameLbl = new ClickableLabel( m_pLayerProp, QSize( 279, 27 ), "",
-										  ClickableLabel::Color::Bright, true );
-	m_pCompoNameLbl->move( 5, 4 );
-	connect( m_pCompoNameLbl, SIGNAL( labelClicked(ClickableLabel*) ),
-			 this, SLOT( renameComponentAction() ) );
-
-	m_DropDownCompoBtn = new Button( m_pLayerProp, QSize( 18, 18 ),
-										Button::Type::Push, "dropdown.svg", "",
-										false, QSize( 12, 12 ) );
-	m_DropDownCompoBtn->setObjectName( "InstrumentEditorComponentNameDropDown" );
-	m_DropDownCompoBtn->move( 263, 8 );
-	connect( m_DropDownCompoBtn, SIGNAL( clicked() ),
-			 this, SLOT( onDropDownCompoClicked() ) );
-
-	// Layer preview
-	m_pLayerPreview = new LayerPreview( nullptr, pPanel );
-
-	m_pLayerScrollArea = new QScrollArea( m_pLayerProp);
-	m_pLayerScrollArea->setFrameShape( QFrame::NoFrame );
-	m_pLayerScrollArea->move( 6, 44 );
-	m_pLayerScrollArea->setVerticalScrollBarPolicy( Qt::ScrollBarAlwaysOff );
-	if ( InstrumentComponent::getMaxLayers() > 16 ) {
-		m_pLayerScrollArea->setVerticalScrollBarPolicy( Qt::ScrollBarAlwaysOn );
-	}
-	m_pLayerScrollArea->setHorizontalScrollBarPolicy( Qt::ScrollBarAlwaysOff );
-	m_pLayerScrollArea->setMaximumHeight( 182 );
-	m_pLayerScrollArea->setWidget( m_pLayerPreview  );
-
-
-	// Waveform display
-	m_pWaveDisplay = new WaveDisplay( m_pLayerProp );
-	m_pWaveDisplay->resize( 277, 58 );
-	m_pWaveDisplay->updateDisplay( nullptr );
-	m_pWaveDisplay->move( 5, 241 );
-	connect( m_pWaveDisplay, SIGNAL( doubleClicked(QWidget*) ),
-			 this, SLOT( waveDisplayDoubleClicked(QWidget*) ) );
-
-	m_pLoadLayerBtn = new Button( m_pLayerProp, QSize( 92, 18 ), Button::Type::Push,
-								  "", pCommonStrings->getLoadLayerButton() );
-	m_pLoadLayerBtn->setObjectName( "LoadLayerButton" );
-	m_pLoadLayerBtn->move( 5, 304 );
-
-	m_pRemoveLayerBtn = new Button( m_pLayerProp, QSize( 94, 18 ), Button::Type::Push,
-									"", pCommonStrings->getDeleteLayerButton() );
-	m_pRemoveLayerBtn->setObjectName( "RemoveLayerButton" );
-	m_pRemoveLayerBtn->move( 97, 304 );
-
-	m_pSampleEditorBtn = new Button( m_pLayerProp, QSize( 92, 18 ), Button::Type::Push,
-									 "", pCommonStrings->getEditLayerButton() );
-	m_pSampleEditorBtn->setObjectName( "SampleEditorButton" );
-	m_pSampleEditorBtn->move( 191, 304 );
-
-	connect( m_pLoadLayerBtn, SIGNAL( clicked() ),
-			 this, SLOT( loadLayerBtnClicked() ) );
-	connect( m_pRemoveLayerBtn, SIGNAL( clicked() ),
-			 this, SLOT( removeLayerButtonClicked() ) );
-	connect( m_pSampleEditorBtn, SIGNAL( clicked() ),
-			 this, SLOT( showSampleEditor() ) );
-	// Layer gain
-	m_pLayerGainLCD = new LCDDisplay( m_pLayerProp, QSize( 36, 16 ), false, false );
-	m_pLayerGainRotary = new Rotary( m_pLayerProp, Rotary::Type::Normal,
-									 tr( "Layer gain" ), false , 0.0, 5.0);
-	m_pLayerGainRotary->setDefaultValue( 1.0 );
-	connect( m_pLayerGainRotary, &Rotary::valueChanged, [&]() {
-		auto pComponent = m_pInstrumentEditorPanel->getInstrument()->getComponent(
-			m_pInstrumentEditorPanel->getSelectedComponent() );
-		if ( pComponent != nullptr ) {
-			auto pLayer = pComponent->getLayer(
-				m_pInstrumentEditorPanel->getSelectedLayer() );
-			if ( pLayer != nullptr ) {
-				pLayer->setGain( m_pLayerGainRotary->getValue() );
-				updateEditor(); // LCD update
-			}
-		}
-	});
-	m_pLayerGainLbl = new ClickableLabel( m_pLayerProp, QSize( 44, 10 ),
-										  pCommonStrings->getLayerGainLabel() );
-	m_pLayerGainLbl->move( 50, 360 );
-
-	m_pCompoGainLCD = new LCDDisplay( m_pLayerProp, QSize( 36, 16 ), false, false );
-	m_pCompoGainRotary = new Rotary( m_pLayerProp, Rotary::Type::Normal,
-									 tr( "Component volume" ), false, 0.0, 5.0 );
-	m_pCompoGainRotary->setDefaultValue ( 1.0 );
-	connect( m_pCompoGainRotary, &Rotary::valueChanged, [&]() {
-		auto pComponent = m_pInstrumentEditorPanel->getInstrument()->getComponent(
-			m_pInstrumentEditorPanel->getSelectedComponent() );
-		if ( pComponent != nullptr ) {
-			pComponent->setGain( m_pCompoGainRotary->getValue() );
-			updateEditor(); // LCD update
-		}
-	});
-	m_pCompoGainLbl = new ClickableLabel( m_pLayerProp, QSize( 44, 10 ),
-										  pCommonStrings->getComponentGainLabel() );
-	m_pCompoGainLbl->move( 147, 360 );
-
-	m_pLayerPitchCoarseLCD = new LCDDisplay( m_pLayerProp, QSize( 28, 16 ), false, false );
-	m_pLayerPitchFineLCD = new LCDDisplay( m_pLayerProp, QSize( 28, 16 ), false, false );
-	m_pLayerPitchLbl = new ClickableLabel( m_pLayerProp, QSize( 45, 10 ),
-										   pCommonStrings->getPitchLabel() );
-	m_pLayerPitchLbl->move( 17, 412 );
-
-	m_pLayerPitchCoarseRotary = new Rotary( m_pLayerProp, Rotary::Type::Center,
-											tr( "Layer pitch (Coarse)" ), true,
-											Instrument::fPitchMin + fFinePitch,
-											Instrument::fPitchMax - fFinePitch );
-	connect( m_pLayerPitchCoarseRotary, &Rotary::valueChanged, [&]() {
-		const float fNewPitch = round( m_pLayerPitchCoarseRotary->getValue() ) +
-			m_pLayerPitchFineRotary->getValue() / 100.0;
-		auto pComponent = m_pInstrumentEditorPanel->getInstrument()->getComponent(
-			m_pInstrumentEditorPanel->getSelectedComponent() );
-		if ( pComponent != nullptr ) {
-			auto pLayer = pComponent->getLayer(
-				m_pInstrumentEditorPanel->getSelectedLayer() );
-			if ( pLayer != nullptr ) {
-				pLayer->setPitch( fNewPitch );
-				updateEditor(); // LCD update
-			}
-		}
-	});
-	m_pLayerPitchCoarseLbl = new ClickableLabel( m_pLayerProp, QSize( 44, 10 ),
-												 pCommonStrings->getPitchCoarseLabel() );
-	m_pLayerPitchCoarseLbl->move( 61, 412 );
-
-	m_pLayerPitchFineRotary = new Rotary( m_pLayerProp, Rotary::Type::Center,
-										  tr( "Layer pitch (Fine)" ), true, -50.0, 50.0 );
-	connect( m_pLayerPitchFineRotary, &Rotary::valueChanged, [&]() {
-		const float fNewPitch = round( m_pLayerPitchCoarseRotary->getValue() ) +
-			m_pLayerPitchFineRotary->getValue() / 100.0;
-		auto pComponent = m_pInstrumentEditorPanel->getInstrument()->getComponent(
-			m_pInstrumentEditorPanel->getSelectedComponent() );
-		if ( pComponent != nullptr ) {
-			auto pLayer = pComponent->getLayer(
-				m_pInstrumentEditorPanel->getSelectedLayer() );
-			if ( pLayer != nullptr ) {
-				pLayer->setPitch( fNewPitch );
-				updateEditor(); // LCD update
-			}
-		}
-	});
-	m_pLayerPitchFineLbl = new ClickableLabel( m_pLayerProp, QSize( 44, 10 ),
-											   pCommonStrings->getPitchFineLabel() );
-	m_pLayerPitchFineLbl->move( 147, 412 );
-
-	m_pLayerGainLCD->move( 53, 343 );
-	m_pLayerGainRotary->move( 94, 341 );
-
-	m_pCompoGainLCD->move( 151, 343 );
-	m_pCompoGainRotary->move( 191, 341 );
-
-	m_pLayerPitchCoarseLCD->move( 70, 393 );
-	m_pLayerPitchCoarseRotary->move( 105, 391 );
-
-	m_pLayerPitchFineLCD->move( 155, 393 );
-	m_pLayerPitchFineRotary->move( 191, 391 );
-
-	m_pSampleSelectionCombo = new LCDCombo(
-		m_pLayerProp, QSize( width() - 76 - 7, 18 ), true );
-	m_pSampleSelectionCombo->move( 76, 432 );
-
-	m_pSampleSelectionCombo->setToolTip( tr( "Select selection algorithm" ) );
-	setupSampleSelectionCombo();
-	connect( m_pSampleSelectionCombo, SIGNAL( activated( int ) ),
-			 this, SLOT( sampleSelectionChanged( int ) ) );
-	m_pSampleSelectionLbl = new ClickableLabel( m_pLayerProp, QSize( 70, 10 ),
-												pCommonStrings->getSampleSelectionLabel() );
-	m_pSampleSelectionLbl->move( 7, 436 );
-
-	// ~ Layer properties
-
-	//component handling
-	m_pComponentMenu = new QMenu( this );
-	populateComponentMenu();
-	// ~ component handling
-
-	showLayers( false );
 	updateEditor();
 }
 
 InstrumentEditor::~InstrumentEditor() {
 }
 
-void InstrumentEditor::updateEditor()
-{
+void InstrumentEditor::updateEditor() {
 	auto pHydrogen = Hydrogen::get_instance();
 	const auto pInstrument = m_pInstrumentEditorPanel->getInstrument();
-	const int nSelectedComponent =
-		m_pInstrumentEditorPanel->getSelectedComponent();
-	const int nSelectedLayer = m_pInstrumentEditorPanel->getSelectedLayer();
 
 	auto pCommonStrings = HydrogenApp::get_instance()->getCommonStrings();
 
-	bool bLayerActive = false;
 	updateActivation();
 
 	if ( pInstrument != nullptr ) {
@@ -696,118 +469,14 @@ void InstrumentEditor::updateEditor()
 									   Event::Trigger::Suppress );
 		m_pHihatMinRangeLCD->setMaximum( pInstrument->getHigherCc() );
 		m_pHihatMaxRangeLCD->setMinimum( pInstrument->getLowerCc() );
-
-		// see instrument.h
-		m_pSampleSelectionCombo->setCurrentIndex(
-			pInstrument->sampleSelectionAlg() );
-
-		const auto pComponent =
-			pInstrument->getComponent( nSelectedComponent );
-		if ( pComponent != nullptr ) {
-			m_pCompoNameLbl->setText( pComponent->getName() );
-			m_pCompoGainLCD->setText(
-				QString( "%1" ).arg( pComponent->getGain(),
-									 -2, 'f', 2, '0' ) );
-			m_pCompoGainRotary->setValue( pComponent->getGain(), false,
-										  Event::Trigger::Suppress );
-			if ( ! m_pCompoGainRotary->getIsActive() ) {
-				m_pCompoGainRotary->setIsActive( true );
-			}
-
-			if ( nSelectedLayer >= 0 ) {
-				const auto pLayer = pComponent->getLayer( nSelectedLayer );
-
-				if ( pLayer != nullptr ) {
-					bLayerActive = true;
-
-					// Layer GAIN
-					m_pLayerGainRotary->setIsActive( true );
-					m_pLayerGainRotary->setValue( pLayer->getGain(), false,
-												  Event::Trigger::Suppress );
-					m_pLayerGainLCD->setText(
-						QString( "%1" ).arg( pLayer->getGain(), -2, 'f', 2, '0' ) );
-
-					// Component GAIN
-					m_pCompoGainRotary->setIsActive( true );
-					m_pCompoGainRotary->setValue( pComponent->getGain(), false,
-												  Event::Trigger::Suppress );
-					m_pCompoGainLCD->setText(
-						QString( "%1" ).arg( pComponent->getGain(), -2, 'f', 2, '0' ) );
-
-					// Layer PITCH
-					//
-					// For most X.5 values we prefer to round the digit before
-					// point up and set the fine value to -0.5. But this is not
-					// possible for the maximum value and we have to ensure not
-					// to introduce sudden jumps in the fine pitch rotary.
-					float fCoarseLayerPitch;
-					if ( ( m_pLayerPitchFineRotary->getValue() == 50 &&
-						   pLayer->getPitch() -
-						   trunc( pLayer->getPitch() ) == 0.5  ) ||
-						 pLayer->getPitch() == Instrument::fPitchMax ) {
-						fCoarseLayerPitch = trunc( pLayer->getPitch() );
-					}
-					else if ( m_pLayerPitchFineRotary->getValue() == -50 &&
-							  trunc( pLayer->getPitch() ) -
-							  pLayer->getPitch() == 0.5 ) {
-						fCoarseLayerPitch = ceil( pLayer->getPitch() );
-					}
-					else {
-						fCoarseLayerPitch = round( pLayer->getPitch() );
-					}
-
-					const float fFineLayerPitch =
-						pLayer->getPitch() - fCoarseLayerPitch;
-					m_pLayerPitchCoarseRotary->setIsActive( true );
-					m_pLayerPitchCoarseRotary->setValue( fCoarseLayerPitch, false,
-												 Event::Trigger::Suppress );
-					m_pLayerPitchFineRotary->setIsActive( true );
-					m_pLayerPitchFineRotary->setValue( fFineLayerPitch * 100, false,
-													   Event::Trigger::Suppress );
-
-					m_pLayerPitchCoarseLCD->setText(
-						QString( "%1" ).arg( (int) fCoarseLayerPitch ) );
-					m_pLayerPitchFineLCD->setText(
-						QString( "%1" ).arg( fFineLayerPitch * 100, 0, 'f', 0 ) );
-
-					m_pRemoveLayerBtn->setIsActive( true );
-					m_pSampleEditorBtn->setIsActive( true );
-
-					m_pWaveDisplay->updateDisplay( pLayer );
-				}
-			}
-		}
 	}
 	else {
 		m_pNameLbl->setText( "" );
 	}
-
-	if ( ! bLayerActive ) {
-		// Layer GAIN
-		m_pLayerGainRotary->setIsActive( false );
-		m_pLayerGainLCD->setText( "" );
-
-		// Layer PITCH
-		m_pLayerPitchCoarseRotary->setIsActive( false );
-		m_pLayerPitchFineRotary->setIsActive( false );
-
-		m_pLayerPitchCoarseLCD->setText( "" );
-		m_pLayerPitchFineLCD->setText( "" );
-
-		m_pRemoveLayerBtn->setIsActive( false );
-		m_pSampleEditorBtn->setIsActive( false );
-
-		m_pWaveDisplay->updateDisplay( nullptr );
-	}
-
-	populateComponentMenu();
-	m_pLayerPreview->update();
 }
 
 void InstrumentEditor::updateActivation() {
 	if ( m_pInstrumentEditorPanel->getInstrument() != nullptr ) {
-		m_pCompoNameLbl->setEnabled( true );
-
 		m_pAttackRotary->setIsActive( true );
 		m_pDecayRotary->setIsActive( true );
 		m_pSustainRotary->setIsActive( true );
@@ -833,25 +502,9 @@ void InstrumentEditor::updateActivation() {
 		m_pHihatGroupLCD->setIsActive( true );
 		m_pHihatMinRangeLCD->setIsActive( true );
 		m_pHihatMaxRangeLCD->setIsActive( true );
-
-		m_pLayerGainRotary->setIsActive( true );
-		m_pLayerPitchCoarseRotary->setIsActive( true );
-		m_pLayerPitchFineRotary->setIsActive( true );
-
-		m_pSampleSelectionCombo->setIsActive( true );
-		setupSampleSelectionCombo();
-
-		m_pLoadLayerBtn->setIsActive( true );
-		m_pRemoveLayerBtn->setIsActive( true );
-		m_pSampleEditorBtn->setIsActive( true );
-		m_pIsStopNoteCheckBox->setEnabled( true );
-
-		m_DropDownCompoBtn->setIsActive( true );
-		m_pCompoGainRotary->setIsActive( true );
 	}
 	else {
 		m_pNameLbl->setEnabled( false );
-		m_pCompoNameLbl->setEnabled( false );
 
 		m_pAttackRotary->setIsActive( false );
 		m_pDecayRotary->setIsActive( false );
@@ -887,340 +540,7 @@ void InstrumentEditor::updateActivation() {
 		m_pHihatMinRangeLCD->clear();
 		m_pHihatMaxRangeLCD->setIsActive( false );
 		m_pHihatMaxRangeLCD->clear();
-
-		m_pLayerGainRotary->setIsActive( false );
-		m_pLayerGainLCD->clear();
-		m_pLayerPitchCoarseRotary->setIsActive( false );
-		m_pLayerPitchFineRotary->setIsActive( false );
-		m_pLayerPitchCoarseLCD->clear();
-		m_pLayerPitchFineLCD->clear();
-
-		m_pSampleSelectionCombo->setIsActive( false );
-		m_pSampleSelectionCombo->clear();
-
-		m_pLoadLayerBtn->setIsActive( false );
-		m_pRemoveLayerBtn->setIsActive( false );
-		m_pSampleEditorBtn->setIsActive( false );
-		m_pIsStopNoteCheckBox->setChecked( false );
-		m_pIsStopNoteCheckBox->setEnabled( false );
-
-		m_DropDownCompoBtn->setIsActive( false );
-		m_pCompoGainRotary->setIsActive( false );
-		m_pCompoGainLCD->clear();
-
-		m_pWaveDisplay->updateDisplay( nullptr );
 	}
-}
-
-void InstrumentEditor::waveDisplayDoubleClicked( QWidget* pRef )
-{		
-	const auto pInstrument = m_pInstrumentEditorPanel->getInstrument();
-	const int nSelectedComponent =
-		m_pInstrumentEditorPanel->getSelectedComponent();
-	const int nSelectedLayer = m_pInstrumentEditorPanel->getSelectedLayer();
-	if ( pInstrument == nullptr ) {
-		return;
-	}
-	
-	auto pCompo = pInstrument->getComponent( nSelectedComponent );
-	if ( pCompo == nullptr ) {
-		return;
-	}
-			
-	auto pLayer = pCompo->getLayer( nSelectedLayer );
-	if ( pLayer != nullptr ) {
-		showSampleEditor();
-	}
-	else {
-		loadLayerBtnClicked();
-	}
-}
-
-void InstrumentEditor::showSampleEditor()
-{
-	const auto pInstrument = m_pInstrumentEditorPanel->getInstrument();
-	const int nSelectedComponent =
-		m_pInstrumentEditorPanel->getSelectedComponent();
-	const int nSelectedLayer = m_pInstrumentEditorPanel->getSelectedLayer();
-	if ( pInstrument == nullptr ) {
-		return;
-	}
-
-	auto pCompo = pInstrument->getComponent( nSelectedComponent );
-	if ( pCompo == nullptr ) {
-		return;
-	}
-
-	auto pLayer = pCompo->getLayer( nSelectedLayer );
-	if ( pLayer == nullptr ) {
-		return;
-	}
-
-	auto pSample = pLayer->getSample();
-	if ( pSample != nullptr ) {
-		HydrogenApp::get_instance()->showSampleEditor(
-			pSample->getFilepath(), nSelectedComponent, nSelectedLayer );
-	}
-}
-
-void InstrumentEditor::removeLayerButtonClicked()
-{
-	const auto pInstrument = m_pInstrumentEditorPanel->getInstrument();
-	const int nSelectedComponent =
-		m_pInstrumentEditorPanel->getSelectedComponent();
-	const int nSelectedLayer = m_pInstrumentEditorPanel->getSelectedLayer();
-	if ( pInstrument == nullptr ) {
-		return;
-	}
-
-	auto pHydrogen = Hydrogen::get_instance();
-
-	auto pCompo = pInstrument->getComponent( nSelectedComponent );
-	if ( pCompo == nullptr ) {
-		return;
-	}
-
-	pHydrogen->getAudioEngine()->lock( RIGHT_HERE );
-
-	pCompo->setLayer( nullptr, nSelectedLayer );
-
-	pHydrogen->getAudioEngine()->unlock();
-
-	pHydrogen->setIsModified( true );
-
-	// Select next loaded layer - if available - in order to allow for a quick
-	// removal of all layers. In case the last layer was removed, the previous
-	// one will be selected.
-	int nNextLayerIndex = 0;
-	int nCount = 0;
-	for ( int n = 0; n < InstrumentComponent::getMaxLayers(); n++ ) {
-		auto pLayer = pCompo->getLayer( n );
-		if ( pLayer != nullptr ){
-			nCount++;
-
-			if ( nNextLayerIndex <= nSelectedLayer &&
-				 n != nSelectedLayer ) {
-				nNextLayerIndex = n;
-			}
-		}
-	}
-
-	m_pInstrumentEditorPanel->setSelectedLayer( nCount );
-	m_pInstrumentEditorPanel->updateEditors();
-}
-
-void InstrumentEditor::loadLayerBtnClicked()
-{
-	const auto pInstrument = m_pInstrumentEditorPanel->getInstrument();
-	const int nSelectedComponent =
-		m_pInstrumentEditorPanel->getSelectedComponent();
-	const int nSelectedLayer = m_pInstrumentEditorPanel->getSelectedLayer();
-	if ( pInstrument == nullptr ) {
-		return;
-	}
-
-	auto pHydrogen = Hydrogen::get_instance();
-
-	QString sPath = Preferences::get_instance()->getLastOpenLayerDirectory();
-	QString sFilename = "";
-	if ( ! Filesystem::dir_readable( sPath, false ) ){
-		sPath = QDir::homePath();
-	}
-
-	// In case the button was pressed while a layer was selected, we
-	// try to use path of the associated sample as default one.
-	if ( nSelectedLayer > 0 ) {
-		auto pComponent = pInstrument->getComponent( nSelectedComponent );
-
-		if ( pComponent != nullptr ) {
-			auto pLayer = pComponent->getLayer( nSelectedLayer );
-
-			if ( pLayer != nullptr ) {
-				auto pSample = pLayer->getSample();
-
-				if ( pSample != nullptr ) {
-					if ( ! pSample->getFilepath().isEmpty() ) {
-						QFileInfo fileInfo( pSample->getFilepath() );
-						sPath = fileInfo.absoluteDir().absolutePath();
-						sFilename = fileInfo.absoluteFilePath();
-					}
-				}
-			}
-		}
-	}
-
-	AudioFileBrowser *pFileBrowser =
-		new AudioFileBrowser( nullptr, true, true, sPath, sFilename );
-	QStringList filename;
-	filename << "false" << "false" << "";
-
-	if (pFileBrowser->exec() == QDialog::Accepted) {
-		filename = pFileBrowser->getSelectedFiles();
-		
-		// Only overwrite the default directory if we didn't start
-		// from an existing file or the final directory differs from
-		// the starting one.
-		if ( sFilename.isEmpty() ||
-			 sPath != pFileBrowser->getSelectedDirectory() ) {
-			Preferences::get_instance()->setLastOpenLayerDirectory(
-				pFileBrowser->getSelectedDirectory() );
-		}
-	}
-
-	delete pFileBrowser;
-
-	if ( filename[2].isEmpty() ) {
-		return;
-	}
-
-	bool fnc = false;
-	if ( filename[0] ==  "true" ){
-		fnc = true;
-	}
-
-	int nLastInsertedLayer = nSelectedLayer;
-	if ( filename.size() > 2 ) {
-		for ( int ii = 2; ii < filename.size(); ++ii ) {
-			int nnLayer = nSelectedLayer + ii - 2;
-			if ( ( ii - 2 >= InstrumentComponent::getMaxLayers() ) ||
-				 ( nnLayer + 1  > InstrumentComponent::getMaxLayers() ) ) {
-				break;
-			}
-
-			auto pNewSample = Sample::load( filename[ii] );
-
-			pHydrogen->getAudioEngine()->lock( RIGHT_HERE );
-
-			// If we're using multiple layers, we start inserting the first
-			// layer at nSelectedLayer and the next layer at nSelectedLayer + 1.
-
-			auto pCompo = pInstrument->getComponent( nSelectedComponent );
-			if ( pCompo == nullptr ) {
-				ERRORLOG( QString( "Selected component [%1] is invalid for the current instrument." )
-						  .arg( nSelectedComponent ) );
-				return;
-			}
-
-			auto pLayer = pCompo->getLayer( nnLayer );
-			if ( pLayer != nullptr ) {
-				// insert new sample from newInstrument, old sample gets deleted
-				// by setSample
-				pLayer->setSample( pNewSample );
-			}
-			else {
-				pLayer = std::make_shared<H2Core::InstrumentLayer>( pNewSample );
-				pCompo->setLayer( pLayer, nnLayer );
-			}
-			nLastInsertedLayer = nnLayer;
-
-			if ( fnc ){
-				QString newFilename = filename[ii].section( '/', -1 );
-				newFilename.replace( "." + newFilename.section( '.', -1 ), "");
-				pInstrument->setName( newFilename );
-			}
-
-			//set automatic velocity
-			if ( filename[1] ==  "true" ){
-				setAutoVelocity();
-			}
-
-			pHydrogen->getAudioEngine()->unlock();
-		}
-
-		pHydrogen->setIsModified( true );
-	}
-
-	m_pInstrumentEditorPanel->setSelectedLayer( nLastInsertedLayer );
-	m_pInstrumentEditorPanel->updateEditors();
-}
-
-
-void InstrumentEditor::setAutoVelocity()
-{
-	const auto pInstrument = m_pInstrumentEditorPanel->getInstrument();
-	const int nSelectedComponent =
-		m_pInstrumentEditorPanel->getSelectedComponent();
-	if ( pInstrument == nullptr ) {
-		return;
-	}
-	auto pCompo = pInstrument->getComponent( nSelectedComponent );
-	if ( pCompo == nullptr ) {
-		return;
-	}
-
-	const int nLayers = pCompo->getLayers().size();
-	if ( nLayers == 0 ) {
-		ERRORLOG( QString( "There are no layers in component [%1]" )
-				  .arg( nSelectedComponent ) );
-		return;
-	}
-
-	const float fVelocityRange = 1.0 / nLayers;
-
-	int nLayer = 0;
-	for ( auto& ppLayer : pCompo->getLayers() ) {
-		if ( ppLayer != nullptr ) {
-			ppLayer->setStartVelocity( nLayer * fVelocityRange );
-			ppLayer->setEndVelocity( nLayer * fVelocityRange + fVelocityRange );
-			
-			++nLayer;
-		}
-	}
-}
-
-void InstrumentEditor::renameComponentAction() {
-	const auto pInstrument = m_pInstrumentEditorPanel->getInstrument();
-	const int nSelectedComponent =
-		m_pInstrumentEditorPanel->getSelectedComponent();
-	if ( pInstrument == nullptr ) {
-		return;
-	}
-
-	const auto pComponent = pInstrument->getComponent( nSelectedComponent );
-	if ( pComponent == nullptr ) {
-		ERRORLOG( QString( "Unable to retrieve selected component [%1]" )
-				  .arg( nSelectedComponent ) );
-		return;
-	}
-
-	auto pHydrogenApp = HydrogenApp::get_instance();
-	const auto pCommonStrings = pHydrogenApp->getCommonStrings();
-
-	const QString sOldName = pComponent->getName();
-	bool bIsOkPressed;
-	const QString sNewName = QInputDialog::getText(
-		this, "Hydrogen", tr( "New component name" ), QLineEdit::Normal,
-		sOldName, &bIsOkPressed );
-
-	if ( bIsOkPressed && sOldName != sNewName ) {
-		 pHydrogenApp->pushUndoCommand(
-			 new SE_renameComponentAction(
-				 sNewName, sOldName, nSelectedComponent ) );
-		 pHydrogenApp->showStatusBarMessage(
-			 QString( "%1: [%2] -> [%3]" )
-					 .arg( pCommonStrings->getActionRenameComponent() )
-					 .arg( sOldName ).arg( sNewName ) );
-	}
-}
-
-void InstrumentEditor::renameComponent( int nComponentId, const QString& sNewName ) {
-	const auto pInstrument = m_pInstrumentEditorPanel->getInstrument();
-	if ( pInstrument == nullptr ) {
-		return;
-	}
-
-	auto pComponent = pInstrument->getComponent( nComponentId );
-	if ( pComponent == nullptr ) {
-		ERRORLOG( QString( "Unable to retrieve component [%1]" )
-				  .arg( nComponentId ) );
-		return;
-	}
-
-	pComponent->setName( sNewName );
-
-	Hydrogen::get_instance()->setIsModified( true );
-
-	updateEditor();
 }
 
 void InstrumentEditor::midiOutChannelChanged( double fValue ) {
@@ -1243,191 +563,4 @@ void InstrumentEditor::midiOutChannelChanged( double fValue ) {
 	}
 
 	m_fPreviousMidiOutChannel = fValue;
-}
-
-void InstrumentEditor::onDropDownCompoClicked() {
-	m_pComponentMenu->popup( m_pCompoNameLbl->mapToGlobal(
-								 QPoint( m_pCompoNameLbl->width() - 40,
-										 m_pCompoNameLbl->height() / 2 ) ) );
-}
-
-void InstrumentEditor::populateComponentMenu() {
-	const auto pInstrument = m_pInstrumentEditorPanel->getInstrument();
-	const int nSelectedComponent =
-		m_pInstrumentEditorPanel->getSelectedComponent();
-	if ( pInstrument == nullptr ) {
-		return;
-	}
-
-	const auto pCommonStrings = HydrogenApp::get_instance()->getCommonStrings();
-
-	m_pComponentMenu->clear();
-
-	// Actions to switch between the drumkits
-	for ( int ii = 0; ii < pInstrument->getComponents()->size(); ++ii ) {
-		const auto ppComponent = pInstrument->getComponent( ii );
-		if ( ppComponent != nullptr ) {
-			auto pAction = m_pComponentMenu->addAction(
-				ppComponent->getName(), this,
-				[=](){ switchComponentAction( ii ); } );;
-			if ( ii == nSelectedComponent ) {
-				m_pComponentMenu->setDefaultAction( pAction );
-			}
-		}
-	}
-	m_pComponentMenu->addSeparator();
-	m_pComponentMenu->addAction( pCommonStrings->getMenuActionAdd(), this,
-								 SLOT( addComponentAction() ) );
-	auto pDeleteAction = m_pComponentMenu->addAction(
-		pCommonStrings->getMenuActionDelete(), this, SLOT( deleteComponentAction() ) );
-	if ( pInstrument->getComponents()->size() < 2 ) {
-		// If there is just a single component present, it must not be removed.
-		pDeleteAction->setEnabled( false );
-	}
-
-	m_pComponentMenu->addAction( pCommonStrings->getMenuActionRename(), this,
-								 SLOT( renameComponentAction() ) );
-}
-
-void InstrumentEditor::addComponentAction() {
-	const auto pInstrument = m_pInstrumentEditorPanel->getInstrument();
-	if ( pInstrument == nullptr ) {
-		return;
-	}
-
-	bool bIsOkPressed;
-	const QString sNewName = QInputDialog::getText(
-		this, "Hydrogen", tr( "Component name" ), QLineEdit::Normal,
-		"New Component", &bIsOkPressed );
-	if ( ! bIsOkPressed ) {
-		// Dialog closed using cancel
-		return;
-	}
-
-	auto pHydrogenApp = HydrogenApp::get_instance();
-	const auto pCommonStrings = pHydrogenApp->getCommonStrings();
-
-	auto pNewInstrument = std::make_shared<Instrument>( pInstrument );
-
-	const auto pNewComponent = std::make_shared<InstrumentComponent>( sNewName );
-	pNewInstrument->addComponent( pNewComponent );
-
-	pHydrogenApp->pushUndoCommand(
-		new SE_replaceInstrumentAction(
-			pNewInstrument, pInstrument,
-			SE_replaceInstrumentAction::Type::AddComponent, sNewName ) );
-	pHydrogenApp->showStatusBarMessage(
-		QString( "%1 [%2]" ).arg( pCommonStrings->getActionAddComponent() )
-		.arg( sNewName ) );
-
-	// New components will be appended and should be selected.
-	m_pInstrumentEditorPanel->setSelectedComponent(
-		pInstrument->getComponents()->size() );
-	m_pInstrumentEditorPanel->updateEditors();
-}
-
-void InstrumentEditor::deleteComponentAction() {
-	const auto pInstrument = m_pInstrumentEditorPanel->getInstrument();
-	const int nSelectedComponent =
-		m_pInstrumentEditorPanel->getSelectedComponent();
-	if ( pInstrument == nullptr ) {
-		return;
-	}
-
-	if ( pInstrument->getComponents()->size() <= 1 ) {
-		ERRORLOG( "There is just a single component remaining. This one can not be deleted." );
-		return;
-	}
-
-	auto pComponent = pInstrument->getComponent( nSelectedComponent );
-	if ( pComponent == nullptr ) {
-		ERRORLOG( QString( "Unable to find selected component [%1]" )
-				  .arg( nSelectedComponent ) );
-		return;
-	}
-	auto pHydrogenApp = HydrogenApp::get_instance();
-	const auto pCommonStrings = pHydrogenApp->getCommonStrings();
-
-	const auto sName = pComponent->getName();
-
-	auto pNewInstrument = std::make_shared<Instrument>( pInstrument );
-	pNewInstrument->removeComponent( nSelectedComponent );
-
-	pHydrogenApp->pushUndoCommand(
-		new SE_replaceInstrumentAction(
-			pNewInstrument, pInstrument,
-			SE_replaceInstrumentAction::Type::DeleteComponent, sName ) );
-	pHydrogenApp->showStatusBarMessage(
-		QString( "%1 [%2]" ).arg( pCommonStrings->getActionDeleteComponent() )
-		.arg( sName ) );
-
-	m_pInstrumentEditorPanel->setSelectedComponent(
-		std::clamp( nSelectedComponent, 0,
-					static_cast<int>(pInstrument->getComponents()->size()) - 2 ) );
-	m_pInstrumentEditorPanel->updateEditors();
-}
-
-void InstrumentEditor::switchComponentAction( int nId ) {
-	const auto pInstrument = m_pInstrumentEditorPanel->getInstrument();
-	if ( pInstrument == nullptr ) {
-		return;
-	}
-
-	const auto pComponent = pInstrument->getComponent( nId );
-	if ( pComponent == nullptr ) {
-		ERRORLOG( QString( "Unable to retrieve component [%1]" )
-				  .arg( nId ) );
-		return;
-	}
-
-	m_pInstrumentEditorPanel->setSelectedComponent( nId );
-	m_pInstrumentEditorPanel->updateEditors();
-}
-
-void InstrumentEditor::sampleSelectionChanged( int selected ) {
-	auto pInstrument = m_pInstrumentEditorPanel->getInstrument();
-	 if ( pInstrument == nullptr ) {
-		 return;
-	 }
-
-	 if ( m_pSampleSelectionCombo->count() < 3 ) {
-		 // We flushed the widget while disabling the InstrumentEditor and are
-		 // in the process of refilling it.
-		 return;
-	 }
-
-	if ( selected == 0 ){
-		pInstrument->setSampleSelectionAlg( Instrument::VELOCITY );
-	}
-	else if ( selected == 1 ){
-		pInstrument->setSampleSelectionAlg( Instrument::ROUND_ROBIN );
-	}
-	else if ( selected == 2){
-		pInstrument->setSampleSelectionAlg( Instrument::RANDOM );
-	}
-}
-
-void InstrumentEditor::showLayers( bool bShow ) {
-	if ( bShow ) {
-		m_pInstrumentProp->hide();
-		m_pShowInstrumentBtn->setChecked( false );
-		m_pLayerProp->show();
-		m_pShowLayersBtn->setChecked( true );
-	}
-	else {
-		m_pInstrumentProp->show();
-		m_pShowInstrumentBtn->setChecked( true );
-		m_pLayerProp->hide();
-		m_pShowLayersBtn->setChecked( false );
-	}
-}
-
-void InstrumentEditor::setupSampleSelectionCombo() {
-	m_pSampleSelectionCombo->clear();
-	/*: Sample selection algorithm available in the instrument editor */
-	m_pSampleSelectionCombo->addItem( tr( "First in Velocity" ) );
-	/*: Sample selection algorithm available in the instrument editor */
-	m_pSampleSelectionCombo->addItem( tr( "Round Robin" ) );
-	/*: Sample selection algorithm available in the instrument editor */
-	m_pSampleSelectionCombo->addItem( tr( "Random" ) );
 }
