+++ src/gui/src/SampleEditor/SampleEditor.cpp
@@ -23,7 +23,8 @@
 #include "SampleEditor.h"
 #include "../HydrogenApp.h"
 #include "../CommonStrings.h"
-#include "../InstrumentEditor/InstrumentEditor.h"
+#include "../InstrumentEditor/ComponentsEditor.h"
+#include "../InstrumentEditor/ComponentView.h"
 #include "../InstrumentEditor/InstrumentEditorPanel.h"
 #include "../InstrumentRack.h"
 
@@ -620,7 +621,8 @@ void SampleEditor::on_PlayPushButton_clicked()
 	}
 
 	const int selectedLayer = HydrogenApp::get_instance()->getInstrumentRack()->
-		getInstrumentEditorPanel()->getSelectedLayer();
+		getInstrumentEditorPanel()->getComponentsEditor()->getCurrentView()->
+		getSelectedLayer();
 
 	std::shared_ptr<Song> pSong = Hydrogen::get_instance()->getSong();
 	if ( pSong == nullptr ) {
@@ -693,7 +695,8 @@ void SampleEditor::on_PlayOrigPushButton_clicked()
 	}
 
 	const int nSelectedlayer = HydrogenApp::get_instance()->getInstrumentRack()->
-		getInstrumentEditorPanel()->getSelectedLayer();
+		getInstrumentEditorPanel()->getComponentsEditor()->getCurrentView()->
+		getSelectedLayer();
 	auto pInstr = pSong->getDrumkit()->getInstruments()->get(
 		pHydrogen->getSelectedInstrumentNumber() );
 	if ( pInstr == nullptr ) {
